backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Leonid Aleksieiev <97108289@st.deltion.nl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12, <=3.13"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
typing-extensions = "^4.12.2"
crewai = "^0.80.0"
flask = "^3.1.0"
langchain-core = "^0.3.18"
pydantic = "^2.9.2"
flask-cors = "^5.0.0"
python-docx = "^1.1.2"
pytz = "^2024.2"
langchain = "^0.3.14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
package-mode = false